feat(rubocop): Add global rubocop config
I use solargraph LSP that uses rubocop by default.  Instead of using the
standardrb plugin for solargraph that can't fix yet, I decided to take
the base.yml configuration that standardrb had and apply it to rubocop.
It is missing some of the standardrb specifics, but it works for now.