feat(vim): Allow FZF to quit on Escape
I have never needed to escape in a FZF window to do something in the
buffer, so instead let Esc quit out of the terminal buffer window.

Why does <Esc><Esc> work?  My theory is that the terminal is being
executed in vim as it's own process (i.e. :edit term://fzf), so when you
hit escape, it will hit escape inside the terminal application and not
inside vim.