add(dotfiles): Updated a lot of stuff
I decided to get rid of the dependency of rcm and instead use my own
shell scripts to perform my linking installs.  While not as automatic,
it is less magical and I have a lot of control over what happens,
allowing my dotfiles to have things that I once used, but no longer do
without much fuss.

I changed my old installer.sh to mac.sh as it is really only used for
mac based installs.  I then put in my setup script a flag to install
those if needed.

I moved folders out of their structural placement into the root to run
scripts on.  Folders that were once in a config folder for rcm now live
in the root, and gives an overview of the items I care or once cared
about.

I also am embracing gruvbox.

The setup entrypoint is in the root.  Other folders have their own setup
that are called from the root setup.  I also split those setups that are
universal (bash, vim, etc.) with those that are mac specific
(hammerspoon, karabiner) to allow me to use this setup in multiple
environments.

I started an installers folder to allow me to install specific pieces of
software and configure them.  The idea is to do this per environment
I use so that I could have the software on any platform.