fix(inputrc): input now does not overwrite prompt
I noticed that when typing off the total columns on the terminal, it was
not going to a new line, but was overwriting the current line with
characters.  It has to do with the color codes not being properly
escaped.  Adding \001 (which is \[) and \002 (\]) helped this since I am
using [ and ].