add(scripts): Add neat playlists for mpv
The idea is there is a list of youtube playlists I like to listen to, so
I created a ~/.playlists file that contains these, as well as my own
personal playlist in ~/.default_playlist.  The scrips below will use
these to randomly select a playlist or play my own.

Map, twitch-formats, and twitch-view are used to watch Twitch without
using a browser.

Dev is a shortcut to my workspace; it should either attach to a running
instance, start it and attach if it exists, or run a new one.