add(nvim): Redo init.lua to be more "mine"
While my last one was pretty nice, I found that the time it took to get
that all going was not really worth the gain, and the moment I added
another plugin, I had to tweak a lot of other things instead of adding
what I needed to get my work done, and moving on with life.  This is
still pretty distributed, but it definitely is a little more
straightforward for nvim.