add(nvim): Replace termguicolor colorscheme and statusline
The version of gruvbox and status line plugins I was using were very
good, but required that termguicolor be turned on.  After seeing the
difference between alacritty and terminal's load time of `vim -u NONE`,
it was aparent that Terminal is the clear winner in load times, and as
such, termguicolor had to be turned off, and so these needed to be
replaced.