add(thoughtbot): Add thoughtbot dotifles to try
Try out thoughtbot's dotfiles again, and create a toggle between
coc.nvim and native LSP.

The reasoning for coc.nvim/native LSP toggling is due to signature help.
For some reason, Solargraph is not registering itself as able to
complete signatures with native LSP, but is correctly working for
coc.nvim.  A small thing, but it makes me wonder what else is native LSP
with lsp-config is not being set for Solargraph.