add(vim): Switch to minpac/vim packages
I ran into an issue with vim-fugitive that did not allow the statusline
to be updated with the git branch until after a :G command was used.
Since I am not using any lazy loading magic with vim-plug, it made sense
to give minpac another try and rely on the vim packages feature, which
then allowed vim-fugitive's FugitiveStatusline to function as intended.

A residual of this change was that CamelCaseMotion's autoload function
to set the leader binding for it could not be called directly inside of
the vimrc.  So, I created an autocommand to do this on VimEnter, which
has fixed the issue.