update(hammerspoon): Use terminal file
I use a file that I can change the contents of so that my computer knows
which terminal I'd like to default to using, located at
~/.config/terminal.  This has a one line word of the program MacOS is to
launch, currently: iTerm2, Alacritty, or Terminal.

This change allows Hammerspoon to use this data to know which terminal
to activate and focus when I want to use a shortcut key.