add(nvim): Add solargraph disclaimer
Solargraph with rubocop using standard causes some bad errors.  Instead
of attempting to figure out the magic of nvim-lsp-installer and
solargraph to add standardrb, we'll be using the lspconfig for
solargraph which respects bundles and local installs (and works).

This is just a message to remind me when I install it because I see it.