add(zsh): Add flag to switch using nvim
I'm still on the fence whether to go all in on neovim or stick with
regular vim.  I like some of the features, but vim is still so
performant for my work, whereas LSP for ruby and others seems to be
slow.  And I've done so much work already without the need for a LSP for
these kind of things, so is the juice worth the squeeze?