fix(playlist): MPV now will load on screen ran from
I noticed that no matter what settings I had in hammerspoon, MPV would
not respect it, and would not allow it to move from the main monitor.  I
found this was due to the geometry which I suspect does some magic after
it's been loaded, thus disallowing Hammerspoon to do what it wants to
do.  Now it starts on the monitor launched from, and still attempts to
go full screen.